+++ CHANGELOG.md
@@ -20,7 +20,7 @@
 ### Documentation
 - Basic README for installing the model and dependancies
 
-## 1.0 (1 March 2025)
+## 1.0 (15 April 2025)
 
 ### Model Capability Development
 - Added ability to remove vessels from pool
@@ -28,4 +28,8 @@
 - Made phase overlap an easily accessible variable (in scenario yaml)
 - Added ability to specify a port downtime between projects (in scenario yaml)
 - Added mooring to floating project installation
-- Ability to add specific routed distance from ports to project
+- Added ability to add specific routed distance from ports to project
+- Changed run_coral.py to pull all runs within a scenarios folder
+- Removed regional ports
+- Changed port method to allow different foundation and turbine port assignments
+- Updated manager to release foundation vessels as soon as foundation phase is complete